#1999 (インストーラのDBクリーンアップ時に Warning が出る)
 * r22261, r22347 をマージ

+++ html/install/index.php
@@ -197,12 +197,14 @@ switch ($mode) {
         }
 
         //マスターデータのキャッシュを削除
-        $cache_dir = '../../data/cache/';
+        $cache_dir = DATA_REALDIR . 'cache/';
         $res_dir = opendir($cache_dir);
         while ($file_name = readdir($res_dir)){
             //dummy以外は削除
             if ($file_name != 'dummy'){
-                unlink($cache_dir . $file_name);
+                if (is_file($cache_dir . $file_name)) {
+                    unlink($cache_dir . $file_name);
+                }
             }
         }
         closedir($res_dir);
